requirements.txt
Flask
openai
python-dotenv
PyJWT